requirements.txt
python-telegram-bot
telethon